+++ server_network/src/main.rs
@@ -2,6 +2,7 @@ mod database;
 mod http_methods_implementation;
 mod information;
 mod query;
+mod pagination;
 
 use crate::database::DatabaseConnection;
 use dotenv::dotenv;
